Fix: add generics.get_object_ to posts/views.py

+++ social_media_api/posts/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render
 
 # Create your views here.
-from rest_framework import viewsets, permissions
+from rest_framework import viewsets, permissions, generics, status
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework.filters import SearchFilter
@@ -57,12 +57,14 @@ class LikePostView(APIView):
     permission_classes = [permissions.IsAuthenticated]
 
     def post(self, request, pk):
-        post = get_object_or_404(Post, pk=pk)
+        post = generics.get_object_or_404(Post, pk=pk)
 
         # Prevent duplicate likes
-        like, created = Like.objects.get_or_create(user=request.user, post=post)
-        if not created:
-            return Response({"detail": "You already liked this post."}, status=status.HTTP_400_BAD_REQUEST)
+        if Like.objects.filter(user=request.user, post=post).exists():
+            return Response({"error": "You already liked this post."}, status=status.HTTP_400_BAD_REQUEST)
+        
+        #create like
+         Like.objects.create(user=request.user, post=post)
 
         # Create a notification for the post author (if not self)
         if post.author != request.user:
@@ -80,11 +82,11 @@ class UnlikePostView(APIView):
     permission_classes = [permissions.IsAuthenticated]
 
     def post(self, request, pk):
-        post = get_object_or_404(Post, pk=pk)
+        post = generics.get_object_or_404(Post, pk=pk)
 
-        try:
-            like = Like.objects.get(user=request.user, post=post)
-            like.delete()
-            return Response({"message": "Post unliked successfully!"}, status=status.HTTP_200_OK)
-        except Like.DoesNotExist:
-            return Response({"detail": "You haven't liked this post."}, status=status.HTTP_400_BAD_REQUEST)
+        like = Like.objects.filter(user=request.user, post=post).first()
+        if not like:
+            return Response({"error": "You have not liked this post."}, status=status.HTTP_400_BAD_REQUEST)
+
+        like.delete()
+        return Response({"message": "Post unliked."}, status=status.HTTP_200_OK)
